refactor: move implementation of store to kvstore package
